Vary and Content-Length are each their own question
Two more clusters out of the header-map module, both carrying the typed enum
that is the reason they are worth naming.

`VaryNomination` has three answers where a `Vec<String>` would have two: a set
of field names, `*`, or nothing said. `*` is not a name and never matches, so a
caller that flattened it to an empty set would read "vary on everything, cache
nothing" as "vary on nothing" — the enum is what makes that unspellable.

`ContentLengthError` separates the three ways the field fails, which are three
findings and want three sentences. Beside it, `content_evidence` answers the
different question of which mechanism frames the body at all, since
`Transfer-Encoding` outranks the field and a body may be framed by close. Those
two had been adjacent long enough to look like one question.

headers.rs: 2344 → 914 non-test lines.
