+++ lint-http-rules/src/helpers/headers.rs
@@ -7,144 +7,6 @@ use crate::helpers::quoted_string::{unescape_quoted_string, validate_quoted_stri
 use crate::helpers::shown::describe_char;
 use hyper::HeaderMap;
 
-/// Errors returned by `validate_content_length`.
-#[derive(Debug, PartialEq, Eq)]
-pub enum ContentLengthError {
-    InvalidCharacter(String),
-    TooLarge(String),
-    MultipleValuesDiffer(String, String),
-    NonUtf8,
-}
-
-/// Validate `Content-Length` headers in a `HeaderMap`.
-///
-/// Checks:
-/// 1. Values must be valid UTF-8 digits.
-/// 2. Values must be parseable as u128.
-/// 3. If multiple values are present, they must be identical.
-///
-/// Returns `Ok(None)` if no Content-Length header is present,
-/// `Ok(Some(n))` if valid, or `Err(ContentLengthError)`.
-pub fn validate_content_length(headers: &HeaderMap) -> Result<Option<u128>, ContentLengthError> {
-    let entries: Vec<_> = headers
-        .get_all(hyper::header::CONTENT_LENGTH)
-        .iter()
-        .collect();
-
-    if entries.is_empty() {
-        return Ok(None);
-    }
-
-    let mut first_val: Option<u128> = None;
-    let mut first_raw: String = String::new();
-
-    for hv in entries.iter() {
-        let s = hv.to_str().map_err(|_| ContentLengthError::NonUtf8)?;
-
-        // A single field line may itself carry a comma-separated list, and that is
-        // not automatically a violation. The sentence below makes `5, 5` valid and
-        // says what it means: one value, 5. Splitting here is what keeps this
-        // function's answer the same for a message however its field lines were
-        // merged -- two `Content-Length: 5` lines and one `Content-Length: 5, 5`
-        // line are the same message by § 5.3, and used to get opposite answers.
-        //
-        // cite(RFC 9112 § 6.3): "If a message is received without Transfer-Encoding and with an invalid Content-Length header field, then the message framing is invalid and the recipient MUST treat it as an unrecoverable error, unless the field value can be successfully parsed as a comma-separated list (Section 5.6.1 of [HTTP]), all values in the list are valid, and all values in the list are the same (in which case, the message is processed with that single value used as the Content-Length field value)."
-        let mut saw_value = false;
-        for t in list_members(s) {
-            saw_value = true;
-
-            // cite(RFC 9110 § 8.6, label: Content-Length grammar): "Content-Length = 1*DIGIT"
-            if !t.chars().all(|c| c.is_ascii_digit()) {
-                return Err(ContentLengthError::InvalidCharacter(s.to_string()));
-            }
-
-            // A digit run too long for `u128` is still valid `1*DIGIT` — the grammar
-            // sets no ceiling and, unlike `delta-seconds`, nothing here tells a
-            // recipient to clamp an unrepresentable value. So this rejection is a
-            // tolerance rather than a quoted requirement, and it is left in place on
-            // the reasoning that a length no recipient can represent cannot frame a
-            // message either: §6.3 makes an unusable Content-Length a framing error.
-            // It takes a 39-digit value to reach.
-            let n = t
-                .parse::<u128>()
-                .map_err(|_| ContentLengthError::TooLarge(s.to_string()))?;
-
-            match first_val {
-                None => {
-                    first_val = Some(n);
-                    first_raw = s.to_string();
-                }
-                Some(f) if f != n => {
-                    return Err(ContentLengthError::MultipleValuesDiffer(
-                        first_raw,
-                        s.to_string(),
-                    ));
-                }
-                _ => {}
-            }
-        }
-
-        // An empty field line has no values to be "all the same", and one is not a
-        // list of none: `Content-Length:` is simply not `1*DIGIT`.
-        if !saw_value {
-            return Err(ContentLengthError::InvalidCharacter(s.to_string()));
-        }
-    }
-
-    Ok(first_val)
-}
-
-/// What a captured message shows of its own content, for the rules whose
-/// sentence is about content rather than about framing.
-#[derive(Debug, PartialEq, Eq)]
-pub enum ContentEvidence {
-    /// Octets that streamed through, counted after framing was removed.
-    Captured(u64),
-    /// No body was captured, so the sender's own declaration is what is left.
-    Declared(u128),
-}
-
-impl std::fmt::Display for ContentEvidence {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ContentEvidence::Captured(n) => write!(f, "{n} octets captured"),
-            ContentEvidence::Declared(n) => write!(f, "Content-Length: {n}"),
-        }
-    }
-}
-
-/// Whether a message carries content, and on what evidence.
-///
-/// Content is the octet stream left once framing has been taken off, so a
-/// framing field is not itself an answer: a chunked message whose only chunk is
-/// the terminator carries none, and over HTTP/2 and HTTP/3 content arrives with
-/// no framing field at all.
-// cite(RFC 9110 § 6.4): "HTTP messages often transfer a complete or partial representation as the message "content": a stream of octets sent after the header section, as delineated by the message framing."
-// cite(RFC 9110 § 6.4): "For example, an HTTP/1.1 message body (Section 6 of [HTTP/1.1]) might consist of a stream of data encoded with the chunked transfer coding -- a sequence of data chunks, one zero-length chunk, and a trailer section -- whereas the content of that same message includes only the data stream after the transfer coding has been decoded; it does not include the chunk lengths, chunked framing syntax, nor the trailer fields (Section 6.5)."
-//
-// The captured count is that stream measured directly, so it answers first; a
-// coded representation of nothing is still nothing, which is why a
-// `Content-Encoding` does not disturb the comparison against zero.
-pub fn content_evidence(headers: &HeaderMap, body_length: Option<u64>) -> Option<ContentEvidence> {
-    match body_length {
-        Some(n) => (n > 0).then_some(ContentEvidence::Captured(n)),
-        None => declared_content_length(headers)
-            .filter(|n| *n > 0)
-            .map(ContentEvidence::Declared),
-    }
-}
-
-/// The sender's own claim about how much content it enclosed, where that claim
-/// parses.
-///
-/// A value that does not parse leaves no number and `content_length_valid`
-/// reports it; a declared length that disagrees with the captured octets is
-/// `request_body_length_accuracy`'s finding.
-// cite(RFC 9110 § 8.6): "When transferring a representation as content, Content-Length refers specifically to the amount of data enclosed so that it can be used to delimit framing"
-pub fn declared_content_length(headers: &HeaderMap) -> Option<u128> {
-    validate_content_length(headers).ok().flatten()
-}
-
 /// Retrieve a header value as a string, if it exists and contains only visible ASCII.
 ///
 /// Returns `None` if the header is missing or contains non-visible ASCII characters
@@ -806,81 +668,6 @@ pub fn is_nominated_by_connection(name: &str, connection_header_value: Option<&s
     list_members(conn).any(|tok| tok.eq_ignore_ascii_case(name_l.as_str()))
 }
 
-/// What a response's `Vary` nominates.
-///
-/// The two arms are the two alternatives of `Vary = #( "*" / field-name )`, and
-/// they are kept apart because a caller cannot treat them the same: `*` says the
-/// response varies on something a request's fields do not name, which RFC 9111
-/// § 4.1 turns into *always fails to match*, while a list of names is something
-/// a cache can compare. A member of `*` anywhere in the list makes the whole
-/// value that, since the sentence says *containing a member*.
-pub enum VaryNomination {
-    /// Some member is `*`.
-    Wildcard,
-    /// The field names nominated, folded to lowercase, in the order written.
-    /// Empty for a field that is absent and for a zero-element list, which
-    /// nominate the same nothing.
-    Fields(Vec<String>),
-}
-
-impl VaryNomination {
-    /// Whether `name` is nominated. `*` nominates every name.
-    pub fn nominates(&self, name: &str) -> bool {
-        match self {
-            Self::Wildcard => true,
-            Self::Fields(fields) => {
-                let name = name.trim().to_ascii_lowercase();
-                fields.contains(&name)
-            }
-        }
-    }
-
-    /// Whether some member is `*`.
-    pub fn is_wildcard(&self) -> bool {
-        matches!(self, Self::Wildcard)
-    }
-}
-
-/// Read a response's `Vary` field.
-///
-/// **Three rules asked this question and the three disagreed; two of them were
-/// wrong, and in the same two ways.** Both read `get_all("vary")` line by line
-/// and bailed on a line `to_str` refused — one returning no finding, the other
-/// silently skipping the line — so a single `obs-text` octet anywhere in the
-/// value stood the whole rule down. And neither joined the lines, so a `*`
-/// written on a second field line was a `*` to the third rule and not to them.
-/// § 5.3 makes the lines one value; the field is a `#` list, which is what
-/// licenses the join.
-///
-/// The walk is [`list_members`] — a naive comma split, `OWS`-trimmed, empty
-/// members dropped — and this is its second caller. **Naive is the right answer
-/// here and it is the grammar that says so**: `Vary`'s members are `"*"` and
-/// `field-name`, and a `field-name` is a `token`, which admits no
-/// `quoted-string`. Where an element cannot hold one, a DQUOTE is a character
-/// the member may not have, and reading it as opening a quoted-string makes the
-/// *next* comma data when there is no quoted-string for it to be data in — so
-/// `Vary: "x, prefer` nominates `Prefer`, and the quote-aware walk one caller
-/// used said it did not.
-///
-/// The fold is licensed: a `field-name` is a field name.
-// cite(RFC 9110 § 12.5.5, label: Vary grammar): "Vary = #( "*" / field-name )"
-// cite(RFC 9110 § 5.1): "Field names are case-insensitive"
-// cite(RFC 9110 § 5.3): "A recipient MAY combine multiple field lines within a field section that have the same field name into one field line, without changing the semantics of the message, by appending each subsequent field line value to the initial field line value in order, separated by a comma (",") and optional whitespace (OWS, defined in Section 5.6.3)."
-// cite(RFC 9111 § 4.1): "A stored response with a Vary header field value containing a member "*" always fails to match."
-pub fn vary_nomination(response_headers: &HeaderMap) -> VaryNomination {
-    let Some(value) = combined_field_value_as_written(response_headers, "vary") else {
-        return VaryNomination::Fields(Vec::new());
-    };
-    let mut fields = Vec::new();
-    for member in list_members(&value) {
-        if member == "*" {
-            return VaryNomination::Wildcard;
-        }
-        fields.push(member.to_ascii_lowercase());
-    }
-    VaryNomination::Fields(fields)
-}
-
 /// The sentence a singleton field says when a message carries more than one of
 /// its field lines.
 ///
@@ -1359,51 +1146,6 @@ mod tests {
         }
     }
 
-    // Quoted-string helper tests
-    /// Two `Content-Length: 5` field lines and one `Content-Length: 5, 5` line are
-    /// the same message -- § 5.3 lets a recipient combine the former into the
-    /// latter -- so they must get the same answer. They used to get opposite ones.
-    #[test]
-    fn content_length_comma_list_matches_repeated_field_lines() {
-        fn one(raw: &str) -> Result<Option<u128>, ContentLengthError> {
-            let mut h = HeaderMap::new();
-            h.append(
-                "content-length",
-                raw.parse::<hyper::header::HeaderValue>().unwrap(),
-            );
-            validate_content_length(&h)
-        }
-        fn many(raws: &[&str]) -> Result<Option<u128>, ContentLengthError> {
-            let mut h = HeaderMap::new();
-            for r in raws {
-                h.append(
-                    "content-length",
-                    r.parse::<hyper::header::HeaderValue>().unwrap(),
-                );
-            }
-            validate_content_length(&h)
-        }
-
-        assert_eq!(one("5, 5"), Ok(Some(5)));
-        assert_eq!(one("5,5"), Ok(Some(5)));
-        assert_eq!(one("5, 5"), many(&["5", "5"]));
-
-        // Differing values are an unrecoverable error either way round.
-        assert!(matches!(
-            one("5, 6"),
-            Err(ContentLengthError::MultipleValuesDiffer(_, _))
-        ));
-        assert!(matches!(
-            many(&["5", "6"]),
-            Err(ContentLengthError::MultipleValuesDiffer(_, _))
-        ));
-
-        // Still not a list of digits.
-        assert!(one("").is_err());
-        assert!(one("5, x").is_err());
-        assert!(one("abc").is_err());
-    }
-
     /// The alternation is decided by the first octet and by nothing else, and
     /// each of the three defects is a distinct answer rather than one `Err`:
     /// the seven callers that used to write this out disagree about `Empty` in
@@ -1509,65 +1251,6 @@ mod tests {
         assert!(validate_ext_value("UTF-8''%ZZ").is_err());
         // Invalid: invalid attr-char
         assert!(validate_ext_value("UTF-8''hello@world").is_err());
-    }
-
-    /// The four axes the three hand copies of this predicate disagreed on: the
-    /// join across field lines, an `obs-text` octet, the walk, and the fold.
-    #[test]
-    fn vary_nomination_reads_the_whole_field() {
-        use hyper::header::{HeaderName, HeaderValue};
-        let read = |lines: &[&[u8]]| {
-            let mut h = HeaderMap::new();
-            for l in lines {
-                h.append(
-                    HeaderName::from_static("vary"),
-                    HeaderValue::from_bytes(l).expect("a test Vary value"),
-                );
-            }
-            vary_nomination(&h)
-        };
-
-        // A `*` written on a second field line is a `*`: §5.3 makes the lines
-        // one value and `#` is what licenses the join. Two of the three rules
-        // read the lines one at a time and said it was not.
-        assert!(read(&[b"accept", b"*"]).is_wildcard());
-        assert!(read(&[b"*"]).is_wildcard());
-        assert!(read(&[b"accept, *, accept-encoding"]).is_wildcard());
-
-        // One `obs-text` octet used to stand the whole rule down, in two rules,
-        // by two different routes. It is a member that nominates nothing and it
-        // does not reach the other members.
-        let n = read(&[b"accept-encoding, caf\xe9", b"*"]);
-        assert!(n.is_wildcard(), "the second line still answers");
-        let n = read(&[b"accept-encoding, caf\xe9"]);
-        assert!(
-            n.nominates("Accept-Encoding"),
-            "the first member still reads"
-        );
-        // The octet stays inside the member it was written in — it is no `OWS`
-        // character and nothing trims it — so that member nominates no field a
-        // request could carry under the name without it.
-        assert!(!n.nominates("caf"), "the octet is part of the member");
-        assert!(
-            n.nominates("caf\u{e9}"),
-            "and the member is read as written"
-        );
-
-        // `field-name = token` admits no `quoted-string`, so a DQUOTE is a
-        // character no member may hold and the comma after it is still a
-        // separator. The quote-aware walk one caller used said otherwise.
-        assert!(read(&[b"\"x, prefer"]).nominates("prefer"));
-
-        // Field names are case-insensitive, both directions.
-        assert!(read(&[b"Accept-Encoding"]).nominates("accept-encoding"));
-        assert!(read(&[b"accept-encoding"]).nominates("ACCEPT-ENCODING"));
-
-        // Absent and zero-element nominate the same nothing, and an empty member
-        // is not an element.
-        assert!(!read(&[]).nominates("accept"));
-        assert!(!read(&[b""]).nominates("accept"));
-        assert!(!read(&[b"a,,b"]).nominates(""));
-        assert!(read(&[b"a,,b"]).nominates("b"));
     }
 
     /// Wire order, the four labels, and the two ways a section is absent — a
