+++ lint-http-rules/src/helpers/mod.rs
@@ -42,6 +42,7 @@ pub mod accept_ranges;
 pub mod auth;
 pub mod cache_control;
 pub mod comment;
+pub mod content_length;
 pub mod content_range;
 pub mod cookie;
 pub mod domain;
@@ -61,4 +62,5 @@ pub mod status;
 pub mod structured_fields;
 pub mod token;
 pub mod uri;
+pub mod vary;
 pub mod websocket;
