+++ lint-http-rules/src/rules/options_method_capabilities.rs
@@ -148,9 +148,10 @@ impl Rule for OptionsMethodCapabilities {
         // `content_type_valid`'s — confirmed by running it, not by
         // reading it.
         // cite(RFC 9110 § 9.3.7): "A client that generates an OPTIONS request containing content MUST send a valid Content-Type header field describing the representation media type."
-        if let Some(evidence) =
-            crate::helpers::headers::content_evidence(&tx.request.headers, tx.request.body_length)
-        {
+        if let Some(evidence) = crate::helpers::content_length::content_evidence(
+            &tx.request.headers,
+            tx.request.body_length,
+        ) {
             if !tx.request.headers.contains_key("content-type") {
                 out.push(self.cited(&RFC_9110_9_3_7, ctx.severity, format!(
                             "OPTIONS request carries content ({evidence}) with no Content-Type header field; RFC 9110 § 9.3.7 says a client that generates an OPTIONS request containing content MUST send a valid Content-Type header field describing the representation media type"
